backend/pyproject.toml
[project]
name = "mi-sensor-collector"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "greenlet>=3.0.0",
    "HAP-python>=4.9.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "bleak>=0.21.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "aiosmtplib>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["F821"]  # SQLAlchemy ORM forward references in Mapped[] type hints
